Save: Fix sign-bit shift for 64-bit values.

+++ src/save.c
@@ -493,7 +493,7 @@ static const char * cyaml__get_double(
  */
 static int64_t cyaml_sign_pad(uint64_t raw, size_t size)
 {
-	uint64_t sign_bit = 1 << (size * CHAR_BIT - 1);
+	uint64_t sign_bit = ((uint64_t)1) << (size * CHAR_BIT - 1);
 	unsigned padding = (sizeof(raw) - size) * CHAR_BIT;
 
 	if ((sign_bit & raw) && (padding != 0)) {
